+++ pkg/session/session_test.go
@@ -16,13 +16,12 @@ import (
 
 func TestNew(t *testing.T) {
 	tests := map[string]struct {
-		client      *http.Client
-		log         log.Interface
-		userAgent   string
-		httpTracing bool
-		expected    *session
+		options  []Option
+		expected *session
+		err      string
 	}{
 		"no options provided, return default session": {
+			options: []Option{WithSigner(&edgegrid.Config{})},
 			expected: &session{
 				client:    http.DefaultClient,
 				signer:    &edgegrid.Config{},
@@ -31,13 +30,42 @@ func TestNew(t *testing.T) {
 				userAgent: "Akamai-Open-Edgegrid-golang/9.0.0 golang/" + strings.TrimPrefix(runtime.Version(), "go"),
 			},
 		},
+		"nil client provided, return error": {
+			options: []Option{WithClient(nil)},
+			err:     "client should not be nil",
+		},
+		"nil log provided, return error": {
+			options: []Option{WithLog(nil)},
+			err:     "logger should not be nil",
+		},
+		"empty user agent provided, return error": {
+			options: []Option{WithUserAgent("")},
+			err:     "user agent should not be empty",
+		},
+		"nil signer provided, return error": {
+			options: []Option{WithSigner(nil)},
+			err:     "signer should not be nil",
+		},
+		"invalid retries provided, return error": {
+			options: []Option{WithRetries(RetryConfig{
+				RetryMax:          -1,
+				RetryWaitMin:      -1,
+				RetryWaitMax:      -2,
+				ExcludedEndpoints: []string{"f:o#[]o"},
+			})},
+			err: "retry configuration failed: maximum number of retries cannot be negative\n" +
+				"minimum retry wait time cannot be negative\n" +
+				"maximum retry wait time cannot be negative\n" +
+				"maximum retry wait time cannot be shorter than minimum retry wait time\n" +
+				"malformed exclude endpoint pattern: syntax error in pattern: f:o#[]o",
+		},
 		"with options provided": {
-			client: &http.Client{
-				Timeout: 500,
-			},
-			log:         log.Log,
-			userAgent:   "test user agent",
-			httpTracing: true,
+			options: []Option{
+				WithSigner(&edgegrid.Config{}),
+				WithClient(&http.Client{Timeout: 500}),
+				WithLog(log.Log),
+				WithUserAgent("test user agent"),
+				WithHTTPTracing(true)},
 			expected: &session{
 				client: &http.Client{
 					Timeout: 500,
@@ -52,23 +80,13 @@ func TestNew(t *testing.T) {
 
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			cfg := &edgegrid.Config{}
-			options := []Option{WithSigner(cfg)}
-			if test.client != nil {
-				options = append(options, WithClient(test.client))
+			res, err := New(test.options...)
+			if test.err != "" {
+				assert.EqualError(t, err, test.err)
+			} else {
+				require.NoError(t, err)
+				assert.Equal(t, test.expected, res)
 			}
-			if test.log != nil {
-				options = append(options, WithLog(test.log))
-			}
-			if test.userAgent != "" {
-				options = append(options, WithUserAgent(test.userAgent))
-			}
-			if test.httpTracing {
-				options = append(options, WithHTTPTracing(test.httpTracing))
-			}
-			res, err := New(options...)
-			require.NoError(t, err)
-			assert.Equal(t, test.expected, res)
 		})
 	}
 }