+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 
 
 * Improved formatting of validation errors
-
+* Added ability to return an error to `session.Option`